// src/main/resources/www2/src/api/api.ts
export const RsTaskRestApiViewScheduleTypeEnum = {
    Asap: 'ASAP',
    At: 'AT',
    Daily: 'DAILY',
    Monthly: 'MONTHLY',
    Weekly: 'WEEKLY',
    Hourly: 'HOURLY'
} as const;

export type RsTaskRestApiViewScheduleTypeEnum = typeof RsTaskRestApiViewScheduleTypeEnum[keyof typeof RsTaskRestApiViewScheduleTypeEnum];
// ...
export interface TaskRestApiView {
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'author'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'changeDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'comments'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'creationDate'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TaskRestApiView
     */
    'debugEnabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'executionDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof TaskRestApiView
     */
    'id'?: number;
    /**
     * 
     * @type {TaskRestApiViewLogs}
     * @memberof TaskRestApiView
     */
    'logs'?: TaskRestApiViewLogs;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'scheduleReference'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'scheduleType'?: TaskRestApiViewScheduleTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof TaskRestApiView
     */
    'scheduleFactor'?: number;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'status'?: TaskRestApiViewStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'target'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'runnerId'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'nextExecutionDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'taskDescription'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TaskRestApiView
     */
    'repeating'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof TaskRestApiView
     */
    'type': string;
}

export const TaskRestApiViewScheduleTypeEnum = {
    Asap: 'ASAP',
    At: 'AT',
    Daily: 'DAILY',
    Monthly: 'MONTHLY',
    Weekly: 'WEEKLY',
    Hourly: 'HOURLY'
} as const;

export type TaskRestApiViewScheduleTypeEnum = typeof TaskRestApiViewScheduleTypeEnum[keyof typeof TaskRestApiViewScheduleTypeEnum];
export const TaskRestApiViewStatusEnum = {
    Cancelled: 'CANCELLED',
    Failure: 'FAILURE',
    New: 'NEW',
    Running: 'RUNNING',
    Scheduled: 'SCHEDULED',
    Success: 'SUCCESS',
    Waiting: 'WAITING'
} as const;

export type TaskRestApiViewStatusEnum = typeof TaskRestApiViewStatusEnum[keyof typeof TaskRestApiViewStatusEnum];

/**
 * 
 * @export
 * @interface TaskRestApiViewLogs
 */
export interface TaskRestApiViewLogs {
    /**
     * 
     * @type {number}
     * @memberof TaskRestApiViewLogs
     */
    'length'?: number;
}